Apply randomizing to dictionary/volcabulary mode Fix a problem with how setting initialize Fix times up dialog show on Main page

+++ JustRemember/App.xaml.cs
@@ -81,7 +81,7 @@ namespace JustRemember
 		private void DetectResolution()
 		{
 			var viw = ApplicationView.GetForCurrentView();
-			var res = new List<double>() { viw.VisibleBounds.Width, viw.VisibleBounds.Height };
+			var res = new List<int>() { (int)viw.VisibleBounds.Width, (int)viw.VisibleBounds.Height };
 			var max = res.Max() / 2;
 			if (max < res.Min())
 			{
